+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Insert into `search_resource` only resources of configured types (e.g. if a
+  search index is configured to index only items, then only items will be
+  inserted into `search_resource` for that search index).
+
+  It makes the table smaller, and the sync process faster.
+
 ## [0.18.0] - 2025-11-25
 
 - Always index resources in a separate process